+++ java/edu/ualr/bittorrent/interfaces/messages/Have.java
@@ -4,5 +4,5 @@ import edu.ualr.bittorrent.interfaces.HasPeer;
 import edu.ualr.bittorrent.interfaces.Message;
 
 public interface Have extends HasPeer, Message {
-  public long getPieceIndex();
+  public int getPieceIndex();
 }
